% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOreg.R
\name{MOreg}
\alias{MOreg}
\title{MOreg: Fits a Linear Regression Model}
\usage{
MOreg(Y, X)
}
\arguments{
\item{Y}{A numeric vector representing the response variable (dependent variable).}

\item{X}{A numeric matrix or data frame containing the predictor variables (independent variables).}
}
\value{
A list containing the following components:
\describe{
  \item{beta}{A numeric vector of estimated regression coefficients.}
  \item{residuals}{A numeric vector of residuals (differences between observed and fitted values).}
  \item{sigma_sq_hat}{Estimated variance of the residuals (error term).}
  \item{std_err}{A numeric vector of the standard errors of the regression coefficients.}
  \item{tval}{A numeric vector of the t-values for the regression coefficients.}
  \item{pval}{A numeric vector of the p-values for the regression coefficients.}
}
}
\description{
This function fits a linear regression model.
}
\examples{
Y <- c(2.5, 6.8, 8.7, 9)
X <- c(15, 25, 30, 29)
model <- MOreg(Y, X)

}
